Update README.md
Hey, I recommended taking out the '#' on the left side since its just the readme file and not the .R file, because the '#' was making your text really big.  I also took out all of the code at the end. It's my understanding that you don't normally put the code you want to run into the readme file, just the explanation of what you did and why.  It's completely up to you though (I wouldn't want you to take my word for it in case im wrong).  Just some thoughts feel free to merge or not.